+++ lib/instrumentation/generic-span.js
@@ -23,7 +23,6 @@ function GenericSpan (agent, ...args) {
 
   this.timestamp = this._timer.start
   this.ended = false
-  this.sync = true
 
   this.outcome = constants.OUTCOME_UNKNOWN
