+++ lib/instrumentation/span.js
@@ -35,6 +35,7 @@ function Span (transaction, name, ...args) {
   this._message = null
   this._stackObj = null
   this._capturedStackTrace = null
+  this.sync = true
   this._startXid = executionAsyncId()
 
   this.transaction = transaction
