+++ lib/instrumentation/transaction.js
@@ -1,6 +1,5 @@
 'use strict'
 
-const { executionAsyncId } = require('async_hooks')
 var util = require('util')
 
 var ObjectIdentityMap = require('object-identity-map')
@@ -36,7 +35,6 @@ function Transaction (agent, name, ...args) {
   this._contextLost = false // TODO: Send this up to the server some how
   this._abortTime = 0
   this._breakdownTimings = new ObjectIdentityMap()
-  this._startXid = executionAsyncId()
 
   this.outcome = constants.OUTCOME_UNKNOWN
 }
@@ -131,7 +129,6 @@ Transaction.prototype.toJSON = function () {
     result: String(this.result),
     sampled: this.sampled,
     context: undefined,
-    sync: this.sync,
     span_count: {
       started: this._builtSpans
     },
@@ -234,9 +231,6 @@ Transaction.prototype.end = function (result, endTime) {
   this._timer.end(endTime)
   this._captureBreakdown(this)
   this.ended = true
-  if (executionAsyncId() !== this._startXid) {
-    this.sync = false
-  }
 
   var trans = this._agent._instrumentation.currentTransaction
